Refactored `AuthzCheckResource` heirarchy into seperate module for all the right reasons. Added SAME_ORG_CHECK config check.
